+++ hipamd/include/hip/amd_detail/hip_api_trace.hpp
@@ -1059,6 +1059,11 @@ typedef hipError_t (*t_hipMemsetD2D32)(hipDeviceptr_t dst, size_t dstPitch, unsi
                                        size_t width, size_t height);
 typedef hipError_t (*t_hipMemsetD2D32Async)(hipDeviceptr_t dst, size_t dstPitch, unsigned int value,
                                             size_t width, size_t height, hipStream_t stream);
+typedef hipError_t (*t_hipStreamSetAttribute)(hipStream_t stream, hipStreamAttrID attr,
+                                 const hipStreamAttrValue *value);
+typedef hipError_t (*t_hipStreamGetAttribute)(hipStream_t stream, hipStreamAttrID attr,
+                                              hipStreamAttrValue *value_out);
+
 // HIP Compiler dispatch table
 struct HipCompilerDispatchTable {
   // HIP_COMPILER_API_TABLE_STEP_VERSION == 0
@@ -1608,6 +1613,8 @@ struct HipDispatchTable {
   t_hipMemsetD2D16Async hipMemsetD2D16Async_fn;
   t_hipMemsetD2D32 hipMemsetD2D32_fn;
   t_hipMemsetD2D32Async hipMemsetD2D32Async_fn;
+  t_hipStreamGetAttribute hipStreamGetAttribute_fn;
+  t_hipStreamSetAttribute hipStreamSetAttribute_fn;
 
   // HIP_RUNTIME_API_TABLE_STEP_VERSION = 14
   // removed HIP_MEMSET_NODE_PARAMS replaced by hipMemsetParams