+++ hipamd/src/hip_api_trace.cpp
@@ -828,7 +828,7 @@ hipError_t hipEventRecordWithFlags(hipEvent_t event, hipStream_t stream, unsigne
 hipError_t hipLaunchKernelExC(const hipLaunchConfig_t* config, const void* fPtr, void** args);
 hipError_t hipDrvLaunchKernelEx(const HIP_LAUNCH_CONFIG* config, hipFunction_t f, void** params,
                                 void** extra);
-hipError_t hipMemGetHandleForAddressRange(void* handle, hipDeviceptr_t dptr, size_t size, 
+hipError_t hipMemGetHandleForAddressRange(void* handle, hipDeviceptr_t dptr, size_t size,
                                           hipMemRangeHandleType handleType,
                                           unsigned long long flags);
 hipError_t hipMemsetD2D8(hipDeviceptr_t dst, size_t dstPitch, unsigned char value, size_t width,
@@ -843,7 +843,10 @@ hipError_t hipMemsetD2D32(hipDeviceptr_t dst, size_t dstPitch, unsigned int valu
                           size_t height);
 hipError_t hipMemsetD2D32Async(hipDeviceptr_t dst, size_t dstPitch, unsigned int value,
                                size_t width, size_t height, hipStream_t stream);
-
+hipError_t hipStreamGetAttribute(hipStream_t stream, hipStreamAttrID attr,
+                                 hipStreamAttrValue *value);
+hipError_t hipStreamSetAttribute(hipStream_t stream, hipStreamAttrID attr,
+                                 const hipStreamAttrValue *value);
 }  // namespace hip
 
 namespace hip {
@@ -1366,6 +1369,8 @@ void UpdateDispatchTable(HipDispatchTable* ptrDispatchTable) {
   ptrDispatchTable->hipMemsetD2D16Async_fn = hip::hipMemsetD2D16Async;
   ptrDispatchTable->hipMemsetD2D32_fn = hip::hipMemsetD2D32;
   ptrDispatchTable->hipMemsetD2D32Async_fn = hip::hipMemsetD2D32Async;
+  ptrDispatchTable->hipStreamGetAttribute_fn = hip::hipStreamGetAttribute;
+  ptrDispatchTable->hipStreamSetAttribute_fn = hip::hipStreamSetAttribute;
 }
 
 #if HIP_ROCPROFILER_REGISTER > 0
@@ -2017,14 +2022,15 @@ HIP_ENFORCE_ABI(HipDispatchTable, hipMemsetD2D16_fn, 480);
 HIP_ENFORCE_ABI(HipDispatchTable, hipMemsetD2D16Async_fn, 481);
 HIP_ENFORCE_ABI(HipDispatchTable, hipMemsetD2D32_fn, 482);
 HIP_ENFORCE_ABI(HipDispatchTable, hipMemsetD2D32Async_fn, 483);
-
+HIP_ENFORCE_ABI(HipDispatchTable, hipStreamGetAttribute_fn, 484);
+HIP_ENFORCE_ABI(HipDispatchTable, hipStreamSetAttribute_fn, 485);
 // if HIP_ENFORCE_ABI entries are added for each new function pointer in the table, the number below
 // will be +1 of the number in the last HIP_ENFORCE_ABI line. E.g.:
 //
 //  HIP_ENFORCE_ABI(<table>, <functor>, 8)
 //
 //  HIP_ENFORCE_ABI_VERSIONING(<table>, 9) <- 8 + 1 = 9
-HIP_ENFORCE_ABI_VERSIONING(HipDispatchTable, 484)
+HIP_ENFORCE_ABI_VERSIONING(HipDispatchTable, 486)
 
 static_assert(HIP_RUNTIME_API_TABLE_MAJOR_VERSION == 0 && HIP_RUNTIME_API_TABLE_STEP_VERSION == 14,
               "If you get this error, add new HIP_ENFORCE_ABI(...) code for the new function "